Add further coverage for String utils.
Use a dialog instead of a fragment & showing & hiding the action bar

This is what I wanted all along, it turns out.

Replace "query" with the more general "task" concept.

Backup to Dropbox

This copies the files where it should. Unfortunately, the Dropbox app
does not watch the folder's contents, and does not handle folders.

This will require further work.

Do not use a folder in Dropbox. It can't cope with folders.