Move doing multiple queries into a separate method
Do some internal state checking to ensure it's safe to do more queries.

It's up to callers to ensure they don't call it at the wrong time. This is
not violating encapsulation because this method is protected, and this
class exposes its busy state to children.

This is intended to allow subsequent changes to call multiple queries
during normal operation.

This may be abandoned in favour of using chained queries.