Split EditScoreFragment
Splits into EditScoreEventGroupFragment, whichcontains multiple
EditScoreEventFragments.

Add an implementation of HealthScore.Modfication that modifies if the row
ID isn't set. Otherwise it's modify() is a no-op. It is not anticipated
that this will be called very often (or at all) because to modify this db
is to modify the list of scores that we track.

Correct broken cast check.

Allow entry to the empty health score creation activity.

Fix some layout issues

Also make the rating bar explicitly modifiable.

Ensure we use the right arguments bundle; Return the query users.