Don't use alarm IDs
Let the TimingModel figure out when to notify next.

Make sure the TimingModel sets the alarm when notification has happened.

Also make sure that TimingModel only sets an alarm when the application is
created. The notification should only ever come as a result of an alarm
exipring, not the application starting.