Replace old event UTC date times with local date times
Use the current locale as a best-guess for previous locales.

All the timings we care about are time of day, not absolute times. Even in
database records, we need to record local times.

For example, normal drowsiness health scores would look like a bout of
insomnia if the user travelled 12 time zones round the world!