Test a real world upgrade scenario, and correct problems with it
Default date & time strings generated by the DB implementation do not
conform to my time schema.

The app should use UTC times for created & modified time stamps, because
these are not used for measurements. The "when" time stamp must be local
because this is used for local day time measurements such as meals and time
-sensitive values such as drowsiness.