Use only basic columns in updates
Ensure we update rows we depend on before an attempted update of this row.

Ensure we update the primary key IDs of foreign keys.

Throw an unsupported exception when asked to update a row consisting of
only primary key columns - such an update is more properly expressed as a
deletion followed by a new insertion.