Begin implementation of a new method of interacting with the DB
This uses Meal events as a guinea pig to try out using the generated code
as a way of interacting with the database. The intention is that each edit
event activity receives an event row from the framework when it is first
shown, and it uses this to load the corresponding mealEvent and meal rows.