Replace more legacy database code; Make transactions closeable
The application now opens a single database connection at creation, and
this connection is held by a singleton for the remainder of the app's
lifetime.

This allows multiple threads to read/write to the same connection without
worrying about writing to an already closed connection.

Since the app is the context for the database connection, this has an
impact on many of the utility and test methods throughout the code that no
longer need to provide a context of their own.